+++ src/cucu/environment.py
@@ -6,6 +6,7 @@ import shutil
 import sys
 import time
 from functools import partial
+import yaml
 
 from cucu import config, init_scenario_hook_variables, logger
 from cucu.config import CONFIG
@@ -174,6 +175,12 @@ def after_scenario(ctx, scenario):
             browser.quit()
 
         ctx.browsers = []
+
+    cucu_config_filepath = os.path.join(
+        ctx.scenario_logs_dir, "cucu.config.yaml"
+    )
+    with open(cucu_config_filepath, "w") as config_file:
+        config_file.write(CONFIG.hide_secrets(yaml.dump(CONFIG)))
 
 
 def before_step(ctx, step):